Move my file ignore function to a helper - aim to move more here later
